requirements.txt
pandas
openpyxl
hdbcli
python-dotenv
ollama
tabulate
pytest
PyYAML